requirements.txt
black==24.3.0
click==8.1.7
contourpy==1.2.1
cycler==0.12.1
fonttools==4.50.0
kiwisolver==1.4.5
matplotlib==3.8.3
mypy-extensions==1.0.0
neat-python==0.92
numpy==1.26.4
packaging==24.0
pathspec==0.12.1
pillow==10.3.0
platformdirs==4.2.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
six==1.16.0
tomli==2.0.1
typing_extensions==4.10.0
